overhauled to use a mixed method for roles
- uses an average between the two ecdf types for the aspects of a role
instead of trying to 'deskew' values. this was specifically a problem
with skills and preferences where large numbers of zeros or low values
are often present.
- additionally to test, role ratings are set to zero if the associated
skills in the role have no ratings. an example of this are nuns/priests
in the masterwork mod. the caste cannot learn skills (skill rate 0) and
so all roles containing skills are set to zero.
- role highest/lowest drawn values (red/black diamonds) are now no
longer an arbitrary number and are the top/bottom 5% of the ratings