more logging, getType offset
currently on OSX the offset within various object's virtual table's
virtual function's GetType() is 0x4 where on windows/linux it's 0x1.
this is applicable in at least 4 occasions where a type is read, and
potentially more cases. it's been added to the dfinstance to test with
to avoid adding redundant entries into the memory layouts which can't be
exported by the scripts anyway.