Set correct file ownership for Dwarf Therapist.ini in OSX
The reason settings aren't saved properly in OSX is because the config
file ends up getting owned by root. Changing the ownership in between
runs doesn't work because QT's default method of writing a file creates
a temp file (which will be owned by root if the settings sync when)
running as root which it then copies over the original file. We can fix
this by simply changing the file ownership to the current UID and GID
after we construct a `QSettings` object, which is when this generally
seems to happen.

Fixes #72.